client/requirement.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
Flask-APScheduler
Flask-SocketIO
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
greenlet==1.0.0
idna==2.10
Mako==1.1.4
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
requests==2.25.1
tzlocal==2.1
urllib3==1.26.4
websocket-client==0.58.0
pymongo
pymongo[srv]

